Work on Gourmet XML import
This involved splitting out photo utilities into their own module to
avoid duplication, also switched to Pillow rather than PIL due to some
bugginess with PIL not working with a BytesIO object.

Also changed the way ingredients are stored so that ungrouped
ingredients are separate and not in an unnamed group, as it might be
desirable to allow unnamed groups and works better with the Gourmet
format.