Fix bug when editing recipe with errors
Removing ingredients didn't work as new rows were added for the errors.
Instead, put the errors in the same row, but create a new table with the
original data and a row for the errors. This means we can't traverse the
DOM by grabbing any rows, we have to get the direct children.